Fix arguments passed to FileEntry.copyTo() and FileEntry.moveTo().
The copyTo() and moveTo() function in FileEntry takes a DirectoryEntry
argument and an optional new name argument. The name argument is used
to specify the file name, which should not contain any directory
components. This CL fixes a failed invocation of copyTo() due to a path
with directory components being passed as the name argument.

The CL also chanegs the code to pass the correct form of callback to
copyTo() and moveTo().

BUG=chromium-os:26962
TEST=TEST=Manually tested copy, move, and rename operations on GData mount point.

Review URL: http://codereview.chromium.org/9694057

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@126669 4ff67af0-8c30-449e-8e8b-ad334ec8d88c