Get the ProfileService object after NOTIFICATION_SESSION_STARTED is sent to Chrome.
Currently InputMethodMenu continues to read/write the /home/chronos/Default/Preferences file, rather than /home/chronos/user/Preferences, even after sign in because InputMethodMenu gets a pointer to Profile by calling ProfileManager::GetDefaultProfile() *before* sign in.  We have to do it after sign in since the GetDefaultProfile() function returns a different instance depending on the user is signed in or not.

This is the same fix as http://crrev.com/125787 .

BUG=118142
TEST=see the bug

Review URL: https://chromiumcodereview.appspot.com/9651023

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@126844 4ff67af0-8c30-449e-8e8b-ad334ec8d88c