Fix file manager to better handle missing roots and scan failure
On device, FileManager should always see at least one root (Downloads), but when doing development on Linux some engineers may not have any local paths setup.  This CL improves error handling in this case.

Also I realized that our directory scan logic is setup to periodically retry (silently dropping the completion callback on the floor) rather than have any hard failure notification.  I don't know when this could happen in practice, but if it did it could break my redraw reduction optimization. Similarly, if a scan is aborted (eg. because the active root is quickly changed and a new scan is started) we may not get the callback.  I don't see any trivial low-risk solutions to this problem, so given where we are in the schedule I decided it was better to remove this optimization for now rather than risk any further regression.

BUG=chromium-os:27241
TEST=

Review URL: https://chromiumcodereview.appspot.com/9701063

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@126968 4ff67af0-8c30-449e-8e8b-ad334ec8d88c