The find bar can move off of its location at the right hand corner of the screen if it is obscuring text that it has found. However, you can get the find bar to be empty and in the middle of the screen if you press the correct key combinations. In the case of the 'Settings' page 1) Type a 2) Type backspace
This patch makes the find bar return to its home location at the right hand corner of the screen when it has no text.

The find bar should also go back to its home location if there are no matches. For instance:
Note, the find bar will stay in the center of the screen if 1) It has moved to the center of the screen to reveal selected text
2) More text is entered such that there is no 'found text'
eg type 'ae' on settings page.

BUG=116286
TEST=FindInPageControllerTest.FindMovesWhenObscuring

Review URL: https://chromiumcodereview.appspot.com/9564044

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@127037 4ff67af0-8c30-449e-8e8b-ad334ec8d88c