Don't Close control_widget_ during teardown
AppNonClientFrameView's destructor was calling Close on its control_widget_
during destruction. This was harmless when the user un-maximized an
extension pop-up window, but was a use-after-free when the user had
moused over the un-maximize control and then closed the browser (e.g. by
ctrl-shift-q on ChromeOS).

BUG=116886
TEST=manual (comment #8 of 116886)
TEST=views_unittests,aura_unittests,aura_shell_unittests 

Review URL: http://codereview.chromium.org/9706040

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@127060 4ff67af0-8c30-449e-8e8b-ad334ec8d88c