chromeos: Do not peek at /home/chronos/user/Downloads
This is a re-land of crrev.com/127030
The original patch was reverted as it broke
ExtensionFileBrowserPrivateApiTest.FileBrowserMount
This version fixed the test expectation.

Use $HOME/Downloads if running on Linux desktop.

Note that gspencer@ tried to solve this problem in a cleaner way:

http://codereview.chromium.org/9455073/

However his patch didn't work. I patched it locally, but mount
points like /media/removable didn't show up in the file manager,
and the renderer crashed when I clicked on the "Open downloads folder"
in chrome://downloads, as he warned me beforehand. Hence I decided to
go with a simpler approach in this patch.

BUG=chromium-os:27859
TEST=confirm that contents in ~/Downloads are shown in the file browser on Linux desktop

Review URL: https://chromiumcodereview.appspot.com/9699107

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@127125 4ff67af0-8c30-449e-8e8b-ad334ec8d88c