Split ExtensionUpdater into two smaller classes.
The new ExtensionDownloader takes care of fetching manifests, parsing them,
and downloading crx files. These are then passed to a Delegate.

The ExtensionUpdater is one such Delegate, and does the same as before.

This makes it possible for other components to download extensions given an ID
and update URL; in particular, the retail-mode AppPack for chromeos.

BUG=chromium-os:25463
TEST=Exactly everything about installing/updating extensions works exactly the same. All tests pass.

Review URL: https://chromiumcodereview.appspot.com/9695019

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@127206 4ff67af0-8c30-449e-8e8b-ad334ec8d88c