Invert web content when Windows color scheme is white-on-black.
Adds a new interface to listen for changes to system color changes,
which can be used in future changelists to fix other widgets that need
to recompute their color schemes when the user toggles high-contrast mode
on and off.

NTP is currently the only place in Chrome code that uses Windows theme
colors inside the web content area; update this to invert these colors
if we're also inverting the whole web content area.

BUG=112944
TEST=Press Alt+Shift+PrtScr on Windows to toggle high-contrast mode on/off.

Review URL: https://chromiumcodereview.appspot.com/9452033

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@127344 4ff67af0-8c30-449e-8e8b-ad334ec8d88c