ash: Fix a crash on login when using the uber-tray.
When logging in, and NOTIFICATION_LOGIN_USER_CHANGED is emitted, the default
profile may not have been created yet. Trying to create the profile causes in a
failure, since it needs to do IO, and doing so from the UI thread is bad. So
instead of asking for the default-profile, listen to
NOTIFICATION_PROFILE_CREATED to know when a profile is created, and then use it
where necessary.

BUG=chromium-os:27998
TEST=none

Review URL: https://chromiumcodereview.appspot.com/9716026

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@127567 4ff67af0-8c30-449e-8e8b-ad334ec8d88c