Separate SetPassphrase into encryption and decryption APIs
As of today, there is just one SetPassphrase API in ProfileSyncService, SyncBackendHost, SyncManager and SyncInternal, and it is used as a general purpose API for setting a passphrase during encryption and decryption. This is bad because the UI is aware of its intent when SetPassphrase is called, and yet, we discard that intent and recompute it via internal state deep inside SyncInternal.

Ideally, we should allow the UI to declare its intent by specifying whether it is trying to encrypt or decrypt by calling the appropriate API. The end goal of this change is to make the UI more responsive to the user when a passphrase is entered for decryption, because we can now test the entered passphrase against the UI layer's local cache of the cryptographer's keys, and give the user immediate feedback without having to display a spinner while the call is sent down to the syncer thread.

This patch does a bunch of things:
- Separates the SetPassphrase API into SetEncryptionPassphrase and SetDecryptionPassphrase.
- Plumbs these two APIs all the way up the stack, from SyncInternal, SyncManager, SyncbackendHost and ProfileSyncService.
- Adds a synchronous check in SetDecryptionPassphrase to see if the user-entered passphrase works on the cached pending keys, and thereby eliminates the need to show a spinner in the UI while decrypting.
- Updates calls from the UI layer and from unit and integration tests to the correct variant of the API.
- Updates the code in sync_setup_flow.cc to immediately show an error when an incorrect passphrase is entered.

BUG=108718, 95269
TEST=unit_tests, sync_unit_tests, sync_integration_tests, manually run through sign in with / without passphrases, try incorrect passphrases (we now expect instant errors)

Review URL: https://chromiumcodereview.appspot.com/9584040

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@127630 4ff67af0-8c30-449e-8e8b-ad334ec8d88c