Don't filter shortcut keys when a tab is opened in app mode (aka: windowed or pop-up mode).
Currently, certain browser shortcut keys (ctrl-w/n/t) are not sent to the page. This is done to avoid issues with loading the page and waiting for a response before being able to handle the important shortcuts.

This change removes that filtering *iff* the page is app mode (right-click and "Open as Window" from the "New Tab" page).

BUG=84332
TEST=none

Review URL: http://codereview.chromium.org/9701108

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@127787 4ff67af0-8c30-449e-8e8b-ad334ec8d88c