Trigger NotifyPassphraseAccepted from the UI thread when a correct passphrase is entered
As of today, when a user enters a passphrase for decryption while
signing in to sync, the passphrase is tested against the cached pending
keys, and then sent down to the syncer thread so that the cryptographer
can use it for decryption. At the end of decryption,
OnPassphraseAccepted is called.

The problem with this is that in the meantime, if the user brings up the
advanced settings dialog, we might unnecessarily prompt the user for a
passphrase even though the correct one was entered.

This patch adds a call to NotifyPassphraseAccepted in
SyncBackendHost::SetDecryptionPassphrase if the correct passphrase is
entered. This gives the UI immediate feedback when a passphrase is
accepted, and will pave the way for a snappier user experience.

BUG=119191
TEST=sync_unit_tests, sync_integration_tests, Enter a passphrase and immediately open the advanced setup dialog


Review URL: http://codereview.chromium.org/9794007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@127895 4ff67af0-8c30-449e-8e8b-ad334ec8d88c