Fix UserManager to not leak the mock managers created for unit tests.
All tests share the singletons and if mocks are leaked state will be leaked
and possibly corrupted between the tests. This CL tries to force every test
that needs a particular UserManager state to use its own mock object and set
it in the wanted state.

BUG=116996
TEST=unit_tests

Review URL: https://chromiumcodereview.appspot.com/9705016

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@127946 4ff67af0-8c30-449e-8e8b-ad334ec8d88c