NaCl: Allow hardware exception handling to be enabled on Linux and Mac
On Linux and Mac, enabling exception handling is just a matter of
passing a flag through to NaClChromeMainStart().

This change also fixes exception handling on x86-32 Windows, which we
broke a while ago when we made the exception handling syscalls
conditionally enabled on the NaCl side.

Exception handling on x86-64 Windows will require further work to hook
up the debug helper process.  Because of this, and because x86-32
Windows is not covered by the Chromium trybots, I'm not enabling
NaCl's exception handling test for Windows yet.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2651
TEST=run_inbrowser_exception_test in nacl_integration


Review URL: http://codereview.chromium.org/9724002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@128040 4ff67af0-8c30-449e-8e8b-ad334ec8d88c