Hopefully fix crash in GDataOperationRegistry::OnOperationFinish.
The stack signature shows that a destroyed Operation is kept in OperationIDMap, which implies duplicated Operation is inserted to the map (otherwise it shouldn't happen, because we delete iff remove-from-map).

It actually can happen when authentification is failed:

UrlFetchOperationBase::OnURLFetchComplete(HTTP_UNAUTHORIZED)
=> re_authenticate_callback_.Run
===> DocumentsService::RetryOperation

This CL remedies the case.

BUG=chromium-os:28124
TEST=unit_tests --gtest_filter='*GData*'

Review URL: https://chromiumcodereview.appspot.com/9817015

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@128049 4ff67af0-8c30-449e-8e8b-ad334ec8d88c