Clean up a few TODO items.
Push UnhandledWheelEvent and GetBrowserAccessibilityManager down to RenderWidgetHostViewPort.

Get rid of OnMessageReceivedForTesting, in favor of RenderViewHostTester::TestOnMessageReceived (which already existed but needed to be converted to a static method so it could be used with regular RenderViewHost instances).

Rename WebContentsTester::pending_rvh() to GetPendingRvh() since it's a virtual method.

Remove top-level alias for content::RenderViewHostTestHarness and switch clients to use the fully-qualified name.

Clean up a couple of other minor TODO items, some had already been done and the comment just needed to be removed; in one case the TODO was just wrong (in dump_accessibility_tree_browsertest.cc) since it's OK for content to depend on content implementations.

Remove test_tab_contents.h from DEPS (seems to have been a merge failure in my last change that left this in).

BUG=98716

Review URL: http://codereview.chromium.org/9838009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@128263 4ff67af0-8c30-449e-8e8b-ad334ec8d88c