gdata: Remove parasites from GDataFileSystem.
GDataUploader, GDataDownloadObserver, and GDataSyncClient, were owned by
GDataFileSystem, just because there was no better owner.

Introduce GDataSystemService, a profile-keyed service, to be the new owner
of these objects. GDataFileSystem is also owned by the new owner, and
becomes a nomral object than a profile-keyed-service.

The motivation of this change is to make GDataFileSystem, which is already
overly complex, a little bit simpler, and make it easier to write tests.

BUG=chromium-os:28238
TEST=run tests, the file manager works as before

Review URL: https://chromiumcodereview.appspot.com/9808033

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@128334 4ff67af0-8c30-449e-8e8b-ad334ec8d88c