Show GData upload progress in the download shelf.
From the user's perspective, a download looks the same regardless of whether the target is a file on local disk or in the cloud. This patch incorporates the completion state and progress of an associated upload into the display of a download item. In particular:

- Delays the completion of the download item until the associated upload is complete.

- Use the progress of the upload as the progress of the download item in the shelf.

BUG=none
TEST=GData downloads show upload progress in the download shelf.


Review URL: http://codereview.chromium.org/9773001

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@128524 4ff67af0-8c30-449e-8e8b-ad334ec8d88c