gdata: Fix WeakPtr failures from GDataFileSystem
The problem was that we had a WeakPtrFactory created on UI thread in
the constructor of GDataFileSystem, and used it on UI and IO threads,
which we should not (i.e. WeakPtrFactory should only be used on the
thread where the factory was created).

The solution is to create separate factories for UI and IO threads,
and use them on the right threads. To make shutdown safe,
ShutdownOnUIThread() is introduced, which does most parts of
shutdown on UI thread, and then the destructor is called on IO
thread to delete the WeakPtrFactory bound to IO thread.
This shutdown sequence is rather tricky, but a commonly used pattern
in network related code in Chrome.

BUG=chromium-os:28292
TEST=play with the file manager on Debug build, and no assertions failed.

Review URL: https://chromiumcodereview.appspot.com/9835043

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@128547 4ff67af0-8c30-449e-8e8b-ad334ec8d88c