[Sync] Improve nigori conflict and rollback handling.
We now ensure we don't change to an old default key when we receive a nigori
update. We also preserve the local explicit passphrase state if we get into
conflict with a nigori node with old encryption keys. Lastly, we write to the
node at the end of each sync cycle toensure it has the current encrypted types
and encryption keys (assuming the cryptographer is ready). This way, we'll
never lose the current encryption state due to a blank/old nigori.

BUG=119207
TEST=

Review URL: https://chromiumcodereview.appspot.com/9815018

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@128567 4ff67af0-8c30-449e-8e8b-ad334ec8d88c