Clear image_urls_ in FaviconHandler::FetchFavicon
So, it seems that FetchFavicon() is used to explicitly set the url, e.g. for a chrome:// page. If iamge_urls_ is not cleared, the fetch will not succeed. Clearing image_urls_ is safe here for the same reason it was safe in OnUpdateFaviconURL() - we check against the front of image_urls_ on any download so by clearing it any pending downloads will be safely and correctly ignored.

BUG=119601
TEST=See issue.


Review URL: http://codereview.chromium.org/9808091

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@128689 4ff67af0-8c30-449e-8e8b-ad334ec8d88c