[Chrome OS ECHO] Add getCouponCode() and getUserConsent() methods
getUserConsent() pops up an confirmation butter bar to get a user's consent for performing an ECHO transcation. It takes a callback with the result of the user's consent.

getCouponCode is currently a partial stub implementation which will subsequently be hooked to an underlying coupon retrieving system. It currently takes the coupon type to retrieve and returns a mock value.

BUG=chromium-os:27855, chromium-os:27853
TEST=
ExtensionOffersTest API test for getCouponCode
Offers component extension exercises both these methods and displays the results.


Review URL: http://codereview.chromium.org/9729008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@128712 4ff67af0-8c30-449e-8e8b-ad334ec8d88c