Fix ConstrainedWindowViews losing input on focus on aura.
The link in the bug brings up a ConstrainedWindowViews widget with two native textfields. Upon Show(), the focused view is set to NativeTextfieldView, but the focused aura window is set to the browser window (due to ClearNativeFocus). Upon a mouse click, root_window_event_filter corrected the focused aura window to be that of the ConstrainedWindowViews one. This widget however, is not a top level widget and does not have a text input client associated with it. From this point on, all key events are dropped (until we focus on a different view, e.g. click on the second textfield or a button).

The fix here:
- explicitly set focus to constrained window window.
- tell rwhva to not handle input events when constrained window is up
- this has the side effect of calling input_method->OnFocus(), InputMethodBridge needs to allow for non top level widget to call this method.

BUG=118679
TEST=manually on linux.


Review URL: http://codereview.chromium.org/9811028

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@128715 4ff67af0-8c30-449e-8e8b-ad334ec8d88c