NaCl: Make hardware exception handling work on 64-bit Windows
In order to support untrusted hardware exception handling on 64-bit
Windows, we need to have a 64-bit process attach as a debugger to the
64-bit NaCl loader process, so we extend the NaCl broker process to do
this.

The new code in nacl_broker_listener.cc for attaching a 64-bit
debugger is similar to the 32-bit code in nacl_process_host.cc.  I
intend to unify the two in a later change.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2651
TEST=run_inbrowser_exception_test in nacl_integration

Review URL: https://chromiumcodereview.appspot.com/9835003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@128764 4ff67af0-8c30-449e-8e8b-ad334ec8d88c