gdata: Fix issue with copying hosted documents into Docs folder.
This CL changes GDataFileSystem to check if a local file is a hosted
document before uploading the file to the remote server. If the file is
a hosted document, it copies the corresponding document on the server
side and skips the upload. Otherwise, it uploads the file as before.

BUG=chromium-os:28388
TEST=Tested the following:
1. Run "unit_tests --gtest_filter='*GData*'"
2. Manually test copying and moving hosted documents from Downloads
   folder to Docs folder in file manager.

Review URL: https://chromiumcodereview.appspot.com/9837087

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@128826 4ff67af0-8c30-449e-8e8b-ad334ec8d88c