Remove per-tab preference machinery.
It's no longer needed, according to mnaganov@, and it wasn't interacting well
with extension-controlled preferences. This is a partial revert of the
following:

http://codereview.chromium.org/8879016/
http://codereview.chromium.org/8747015/
http://codereview.chromium.org/8716004/
http://codereview.chromium.org/8568019/

The "global" pref names are are kept as is. I plan to rename them back
to the original names after this patch.

Also, per-tab preferences introduced in the following are reverted:
https://chromiumcodereview.appspot.com/9264071/
https://chromiumcodereview.appspot.com/9255018/



BUG=105537
TEST=manually and unit_tests --gtest_filter=Pref* and
browser_tests --gtest_filter=PrefsTabHelperBrowserTest* still pass


Review URL: http://codereview.chromium.org/9838050

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@128871 4ff67af0-8c30-449e-8e8b-ad334ec8d88c