Add reporting notification to the enterprise banner on the CrOS login screen.
This is only shown if there are any reporting features enabled via device policy. A help link takes the user to a helpapp article that explains more about reporting. With this CL, the helpapp link is actually stale, the helpapp change is pending though.

BUG=chromium-os:26625
TEST=Flip reporting policy switches, observe that the enterprise notification indicates reporting is enabled.

Review URL: https://chromiumcodereview.appspot.com/9836039

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@128957 4ff67af0-8c30-449e-8e8b-ad334ec8d88c