Make Mac panels to activate on mousedown in detached mode.
This is to make behavior of detached panels more in line with user expectations on a free-floating windows. It also prevents the drag of non-activated panels across other windows, with a panel getting behind of it and invisible.

XIB change is to change the type of overlay view on top of the Titlebar.

BUG=120340

Review URL: https://chromiumcodereview.appspot.com/9836101

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@129142 4ff67af0-8c30-449e-8e8b-ad334ec8d88c