Aura: Add non-browser windows into the list of "Alt + Tab" cycle.
This CL changes to use the child window list of the default container as the list of "Alt + Tab" cycle, instead of the browsers' list. But the list excludes the window which is not forcusable and has modal window as a child.

This CL can enable us to focus task manager window with "Alt + Tab" switcher.

BUG=118611
TEST=manual

Review URL: http://codereview.chromium.org/9719037

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@129443 4ff67af0-8c30-449e-8e8b-ad334ec8d88c