Optimized fetching of the very first feed
Changed the very first feed from either disk cache or the server to be 200 items only. The remaining chunk of the content will be pulled separately and UI will be receive event notification when it completes. This should significantly speed up startup time for users with 10,000+ items.

BUG=chromium-os:28532
TEST=Unit:BrowserUrlUtilTest.AppendOrReplaceQueryParameter , Manual: make sure first chunk of root directory content shows quickly for any size of docs corpus.

Review URL: https://chromiumcodereview.appspot.com/9918024

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@129711 4ff67af0-8c30-449e-8e8b-ad334ec8d88c