[Sync] Automatically clear passphrase required reason for first time sync
In SetEncryptionPassphrase, if the passphrase required reason is
REASON_ENCRYPTION (and we're not attempting to do an invalid passphrase
transition) we know the passphrase will be accepted. As a result we can
automatically clear the passphrase required reason before the
OnPassphraseAccepted call gets sent.

BUG=121190
TEST=manual first time sync


Review URL: http://codereview.chromium.org/9968020

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@130198 4ff67af0-8c30-449e-8e8b-ad334ec8d88c