Enables a SPDY proxy to push resources from arbitrary origins.
This change will make it possible for Chromium to accept a resource
that is pushed by a SPDY proxy, even if the SYN control frame advertises
a URL whose origin is different from its associated request. The advantage
of this change is, potentially, better performance. This change is disabled
by default and activated by a new switch.

BUG=113427
TEST=


Review URL: http://codereview.chromium.org/9475037

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@130302 4ff67af0-8c30-449e-8e8b-ad334ec8d88c