sync: Don't sound the INVALID_GAIA_CREDENTIALS alarm if credentials are available.
Before this patch, ProfileSyncService would drop into an invalid gaia auth error state whenever it received NOTIFICATION_TOKEN_REQUEST_FAILED. That notification is fired even if a valid token exists in the TokenService and sync was (potentially) authenticated.

This patch takes a more conservative approach where we don't UpdateAuthErrorState if the token service has a token for chromiumsync, and rely on actual auth errors from the SyncBackend to tell us if that token is valid or not.

BUG=121913, 120479
TEST=verifying error messaging when reauthenticating / signing in / restarting while signed in.

Review URL: https://chromiumcodereview.appspot.com/9980020

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@130836 4ff67af0-8c30-449e-8e8b-ad334ec8d88c