ui: Move gesture-recognizer into ui/base/.
Moving the gesture-recognizer into ui/base/ makes it possible to be used by both
aura and views independently. This is the first step: add an interface for
touch-events and gesture-events so that the gesture recognizer does not have to
depend on aura or views events.

BUG=121744
TEST=none

Review URL: https://chromiumcodereview.appspot.com/9958136

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@131023 4ff67af0-8c30-449e-8e8b-ad334ec8d88c