GTK Panels fix for height glitch after opening.
Auto-resizable panels expanded slightly in height when mouse is
moved over a UI element in the panel after first creation and
when re-activated from overflow. This is because the frame
size is reduced if the panel is sized smaller than the frame,
such as the initial minimum size of 100x20 for auto-resizable
panels, or the overflow panel size.

Resolved by storing the frame size from when it was first allocated.

BUG=None
TEST=Open an auto-resizable panel and mouse over its content. No height change.

Review URL: https://chromiumcodereview.appspot.com/9784015

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@131208 4ff67af0-8c30-449e-8e8b-ad334ec8d88c