Have only toplevel functions delete this.
Previous we would "delete this" randomly and possible have use-after-free. This change tries to push all the "delete this" to toplevel functions and return error codes internally. This is still a bit confusing and suboptimal, but should be an improvement and at least some theoretical bugs are fixed with this.

I also changed most of the non-debug logging to DLOG since these bloat the binary and aren't normally useful. I kept a few that seemed more useful, especially about talking to GDB.

BUG=
TEST=

Review URL: https://chromiumcodereview.appspot.com/10007046

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@131576 4ff67af0-8c30-449e-8e8b-ad334ec8d88c