Removed 'splice' listeners in list.
It appears to me that the 'splice' events are redundant because they
are only generated in the ArrayListDataModel.splice() method, which
also generates a 'permuted' event.  Receiving both events causes the
list to be redraw()'n twice.  Also, I don't think that list's
handleDataModelChange_ method is suitable for handling the 'splice'
event because it only invalidates one cachedItems_, ie.
      delete this.cachedItems_[e.index];
But the 'splice' event could represent changes to multiple rows.
BUG=105181
TEST=


Review URL: http://codereview.chromium.org/9816020

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@132051 4ff67af0-8c30-449e-8e8b-ad334ec8d88c