Add an accessibility mode for editable text fields only.
This refactors the triggering of accessibility mode from a boolean state and
command-line flag to an AccessibilityMode enum that's passed directly
when a new RenderView is created, which is also more readable and avoids the
extra command-line flag.

Adds a new mode that only syncs editable text nodes, and enables this on
Windows 8 so that we can support showing the virtual keyboard when an
editable text field gets focus.

BUG=122061,118641
TEST=Adds new test


Review URL: http://codereview.chromium.org/9939011

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@132267 4ff67af0-8c30-449e-8e8b-ad334ec8d88c