Move full-screen detection logic from PanelManager to DisplaySettingsProvider.
Also fixed a bug that pnels remain visible with full screen apps on Snow Leopard.
The fix is to also call orderBack/orderFront when full-screen is detected. This is because the full-sreen window is still in normal level and our panel windows stay in the top of z-order when its level is changed to normal.

BUG=123058
TEST=Manual test by launching panels and make an app full-screen


Committed: http://src.chromium.org/viewvc/chrome?view=rev&revision=132144

Review URL: http://codereview.chromium.org/10051020

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@132487 4ff67af0-8c30-449e-8e8b-ad334ec8d88c