[sync] Fix malformed JSON bug in chrome://sync due to failed datatypes
If a failed datatype error occurs during sync and you navigate to
chrome://sync, the failed datatype error message contains an unnecessary
new line character in it.

This violates the JSON format, and results in a parse error while trying
to process the data on the sync internals page.

This patch removes the new line character from the failed datatypes
error message and makes sure the json in chrome://sync is valid.

BUG=124017
TEST=Open chrome://sync when there is a failed datatype error; ensure no new lines

Review URL: https://chromiumcodereview.appspot.com/10115030

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@132807 4ff67af0-8c30-449e-8e8b-ad334ec8d88c