Attempt to fix use after free reported by valgrind
When on_io_completed_ is signaled on file thread while
GDataFileSystem is waiting for the event in the destructor,
GDataFileSystem could free num_pending_tasks_lock_ before
the lock is released on the file thread.

Also we can avoid locking num_pending_tasks_lock_ in the
destructor if we set initial state of on_io_completed_ to be
signaled.

BUG=122717
TEST=trybots

Review URL: https://chromiumcodereview.appspot.com/10001015

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@133032 4ff67af0-8c30-449e-8e8b-ad334ec8d88c