ignore WebUISend messages from old pages
When the page navigates (or reloads) and a RenderView is reused, in-flight WebUISend messages may still be handled rather than being cancelled. This results in a WebUI receiving unexpected messages (e.g. before initialization has had a chance to take place). This patch makes sure in-flight messages are ignored (in-flight messages will come in before the new document element is ready, thus they will be ignored). I've verified that chrome.send executed before load events are dispatched still works (adding a browsertest for that is a TODO).

BUG=121741
TEST=mashing reload and clicking on an entry in the search engine manager does not cause a crash

Review URL: https://chromiumcodereview.appspot.com/9968123

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@133077 4ff67af0-8c30-449e-8e8b-ad334ec8d88c