[Sync] Keep synced tab favicons in memory instead of in the FaviconService.
Due to the favicon service, history backend, and thumbnail db not being
prepared to expire favicons added via sync, we now keep synced tab favicons
in memory within the SessionModelAssocaitor. We do simple deduping based
on the favicon's url, and provide a GetSyncedFaviconForPageURL method that
fills a string with the bytes of a png-encoded favicon for the url.

BUG=122890
TEST=unit_tests --gtest_filter="*SessionModelAssociator*"


Review URL: http://codereview.chromium.org/10071028

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@133100 4ff67af0-8c30-449e-8e8b-ad334ec8d88c