Resolve the conflict that auto-launch has with the background mode feature.
The Auto-launch experiment writes to the Run Registry key on Windows asking for Chrome to be started at login.
The BackgroundModeManager writes to another Run Registry key asking for Chrome to be started without a window.

We need to consolidate this into one Run key that specifies which feature you want as a command line flag to chrome.exe and make sure if a window is requested, then a window is shown (BackgroundModeManager doesn't care if the window is shown or not).

BUG=123139, 53600
TEST=QA has testing instructions on how to test the experimental auto-launch feature (requires branded build). Not sure about testing instructions for background mode.
Review URL: https://chromiumcodereview.appspot.com/9972012

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@133166 4ff67af0-8c30-449e-8e8b-ad334ec8d88c