NaCl: Don't pass the NaCl process's Windows handle to the renderer process
The process handle is not used since we switched NaCl to using
Chromium's BrokerDuplicateHandle() function.

I have changed NaCl's SelLdrLauncher to not use the process ID or
process handle in this context, so returning dummy values of 0 and -1
will be OK: they will not get passed to kill(), waitpid() or
TerminateProcess().

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2719
TEST=nacl_integration etc.


Review URL: http://codereview.chromium.org/10134029

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@133523 4ff67af0-8c30-449e-8e8b-ad334ec8d88c