Transmit a X-Chrome-UMA-Enabled bit to Google domains from clients that have UMA enabled.
This requires a change to the ChromeResourceDispatcherHostDelegate and profiles where we feed the incognito and metrics-enabled states into the object at creation time, so we can check that field when doing our header setting.

This was originally committed as r134625, but reverted due to CrOS failures, which are now fixed in this patch (see delta from patch sets 1 to 2).

BUG=123609
TEST=With UMA enabled (not Chromium), visit www.google.com and ensure that the request header includes X-Chrome-UMA-Enabled with value "1". Ensure that disabling UMA also disables the transmission of this header entirely. Also, ensure that unit_tests GoogleUtilTests all pass.
TBR=pkasting,ivankr

Review URL: http://codereview.chromium.org/10273028

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@134902 4ff67af0-8c30-449e-8e8b-ad334ec8d88c