Make the Geoposition helper struct public
This CL moves the Geoposition struct from content/common to public/content/common.h so that it can be shared by content and browser. The struct is placed in the |content| namespace and as required by the style guide, all methods except the constructor and Validate() are removed.
This paves the way for a follow-up CL in which the content will provide geolocation information to the browser.

BUG=chromium-os:22036
TEST=Chrome and tests build, run

Review URL: http://codereview.chromium.org/10316007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@134989 4ff67af0-8c30-449e-8e8b-ad334ec8d88c