Use the asynchronous version CopyFromBackingStore in AeroPeekManager
I plan to consolidate RenderWidgetHost::CopyFromBackingStore and RenderWidgetHost::AsyncCopyFromBackingStore.
This is a preparation for that work.

AeroPeekManager is not used by default. It is used only when Chrome is launched with "--enable-aero-peek-tabs" option on Windows 7 with Aero enabled in non-application mode.
This change uses WaitableEvent::Wait to adapt AsyncCopyFromBackingStore for the synchronous usage. I'd like to add suppression for the thread restriction for now in order to make the consolidation work progress.

BUG=120003
TEST=Manual

Review URL: http://codereview.chromium.org/10303002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@135314 4ff67af0-8c30-449e-8e8b-ad334ec8d88c