Save Page As MHTML
--save-page-as-mhtml disables saving pages as html-only and complete with sub-resources in a sibling directory, and enables only saving pages as complete with sub-resources in a single MHTML (MIME-HTML) text file.

The SavePackage class supports all three options, wich will allow saving complete pages to gdata, and may simplify file management for users.

If few enough users need the many-files option, we may be able to remove it and most/all of the SavePackage system.

The generated MHTML is viewable in Chrome on all 4 OSs and Opera (at least on windows).
The generated MHTML is not viewable in chrome when loaded via http[s], only via file://. http://code.google.com/p/chromium-os/issues/detail?id=28654
Chrome recognizes both '.mht' and '.mhtml' as MHTML, however '.mhtml' is more google-able, so that is the default extension when saving MHTML files. Chrome does not sniff '.html' files to see if they are actually MHTML, so it will incorrectly render such mis-named MHTML files.

It is not viewable in Firefox. Firefox tries to help you either open it in another program or download it. There are extensions that provide support for MHTML: MAFF, and UnMHT are two.

It is not viewable in Safari. Safari does not recognize "mht" or "mhtml" as extensions that it can render. Changing the extension to "html" allows Safari to open it, but Safari does not seem to understand the file format so it's rendered incorrectly. A port of UnMHT is available for Safari.

It is not viewable in IE: The webpage cannot be displayed. Most likely cause: •Some content or files on this webpage require a program that you don't have installed.
IE supports MHTML, so it may be possible to tweak the mhtml generator to appease both Opera and IE. http://crbug.com/125477

Regarding OOMs: about:memory showed no change when I saved cnn.com as mhtml (1.4MB), so if there was a spike, it either lasted <200ms or was lost in the noise.

BUG=120416

Review URL: http://codereview.chromium.org/10069014

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@135390 4ff67af0-8c30-449e-8e8b-ad334ec8d88c