New Windows high-contrast mode bubble.
When Windows high-contrast mode is on, pop up new bubble with links to a
high-contrast extension, a search for dark themes, a link to Learn More,
and a Close link to close the bubble. Only shows the first time, as before.

The pref is renamed because the previous behavior was reverted, and users
who got the previous bubble on dev/beta should see this one now.

BUG=112944
TEST=manual testing on Windows: Alt+Shift+PrtScr
TBR=ben


Review URL: http://codereview.chromium.org/10332016

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@135591 4ff67af0-8c30-449e-8e8b-ad334ec8d88c