Disable the error for missing @this annotations
As we noticed in http://codereview.chromium.org/10356042/diff/4001/chrome/browser/resources/chromeos/image_burner.js?column_width=80 we're getting a lot of false positives for this particular error. It's rare that we refer to 'this' in a function that's not a method on a prototype, or a constructor, so I think it's best to just disable this check, so that we don't get a bunch of false positives.

See also: http://code.google.com/p/closure-linter/issues/detail?id=43

BUG=none
TEST=none
NOTRY=true

Review URL: http://codereview.chromium.org/10380035

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@135711 4ff67af0-8c30-449e-8e8b-ad334ec8d88c