cros: Fix theme window frames to not paint transparent
We were painting the theme frame image with transparency, which made it almost invisible in solo-window mode, then painting the theme overlay fully opaque.  This lead to incorrect rendering of many themes.  We now follow the pattern we use for Windows Aero glass - all theme images are drawn fully opaque.

Also, some themes don't supply inactive versions of their theme_frame image, and rely on the theme pack to generate one from the active image.  This interfered with a hack I was doing to override the IDR_THEME_FRAME_INACTIVE ID where there wasn't a custom image supplied.  I got rid of the hack and we now use the canonical IDR_THEME_FRAME_* resource IDs on Ash and I updated the default assets the version we're using.

We should probably backport this to R20, since it's a regression in theme painting and the fix is small and safe.

BUG=124095
TEST=Added FramePainterTest.GetHeaderOpacity().
R=derat@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10310065

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@135951 4ff67af0-8c30-449e-8e8b-ad334ec8d88c