Move Declarative Web Request API out of experimental and make it a feature
This CL moves the declarative web request API and the declarative API out of the experimental namespace. Instead the APIs become features that are only enabled on the dev channel.

This CL introduces also empty permissions for the APIs, we need to improve this so that host permissions are respected.

BUG=112155
TEST=no

Review URL: https://chromiumcodereview.appspot.com/10257006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@136006 4ff67af0-8c30-449e-8e8b-ad334ec8d88c