Some minor modifications to the UI for debugging the omnibox ranking.
We are not using a borderbox any more. We are using different shades of gray for the background color instead.

URLs that use to appear as a text are now URL links.

Do not allow wrapping on the first table line where the descriptions of all the columns are.

Rearranging the order of some columns.

BUG=
TEST=


Review URL: https://chromiumcodereview.appspot.com/9956034

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@136062 4ff67af0-8c30-449e-8e8b-ad334ec8d88c