gdata: fix problem with periodic updates when network connection is interrupted.
GDataFileSystem::CheckForUpdates sets the content origin of root directory to
REFRESHING and expects OnFeedFromServerLoaded to set it to FROM_SERVER upon
successful reload of feed. However when we lose network connection, the root
origin is stuck at the REFRESHING state, preventing future updates.

Fix is to implement a callback that checks for failures in reload of feed,
and restores root origin to the initial value when that happens.

BUG=127245
TEST=locally on lumpy device

Review URL: https://chromiumcodereview.appspot.com/10383080
Patch from Haixia Shi <hshi@chromium.org>.

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@136067 4ff67af0-8c30-449e-8e8b-ad334ec8d88c