gdata: Replace "gdata" with "drive" for everything visible in UI.
Note: The scope of this CL is limited to the replacement of
"gdata" with "drive" in string literals that may appear in
URL paths. It is not intended to rename Javascript/C++
namespaces, attributes, enums, structs and classes, or
to relocate header files, source files and test data files.

BUG=126985
TEST=locally tested on lumpy, plus passed try tests

Review URL: https://chromiumcodereview.appspot.com/10386058
Patch from Haixia Shi <hshi@chromium.org>.

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@136257 4ff67af0-8c30-449e-8e8b-ad334ec8d88c