Give browser process the ability to enable and disable NaCl's validation cache.
This is groundwork for persisting the cache to disk. The browser needs to know
when the cache is disabled so that it can avoid reading it off disk.  This will
help keep work on the cache from affecting Chrome before the cache is enabled.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=2515
TEST= none

Review URL: https://chromiumcodereview.appspot.com/10377156

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@137329 4ff67af0-8c30-449e-8e8b-ad334ec8d88c