Revert 137388 - Move declarative API into events API
The motivation of this CL is to bring the declarative API and the rest of the
events mechanism closer together in the API documentation. To achive this, we
move the content of declarative.json and everything related to events from
extension.json into the events.json API definition. The new "events" namespace
is automatically provided to all extensions.

BUG=126497,112155
TEST=no

Review URL: https://chromiumcodereview.appspot.com/10392008

TBR=battre@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10389168

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@137394 4ff67af0-8c30-449e-8e8b-ad334ec8d88c