Ensure that TemplateURLs removed by prepopulate search engine merging is also removed from Sync.
We do this by keeping a list of pre-sync deleted GUIDs, which is populated whenever a TemplateURL entry is removed from the DB during startup merging. We then use this list to ignore the appropriate entries from Sync, and to inform Sync that they should be deleted.

BUG=128210
TEST=unit_test TemplateURLServiceSyncTest.PreSyncDeletes


Review URL: https://chromiumcodereview.appspot.com/10384188

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@137490 4ff67af0-8c30-449e-8e8b-ad334ec8d88c