Only return the visible page actions from PageActionController.
This is a bit of a yak shave to support having the "action box" show active
extensions on the current page rather than page actions. Previously,
PageActionController was creating widgets for everything then
LocationBarViewGtk was hiding them as necessary. This is assumes too much; we
need to push the visibility logic into the Controller for the future.

BUG=127988


Review URL: https://chromiumcodereview.appspot.com/10388160

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@137630 4ff67af0-8c30-449e-8e8b-ad334ec8d88c