cros: Fix focus not being released from omnibox when entering fullscreen
This would result in typed text spawning the omnibox results on top of the page.  Focus needed to be released whenever the location bar view, or any child, had focus, as the omnibox is constructed slightly differently on Chrome OS.  Also added a test for this condition.  I verified by hand this still works on Windows.

BUG=125204
TEST=browser_tests BrowserViewTest.* also manually enter fullscreen and verified typing "www" no longer spawns omnibox results


Review URL: https://chromiumcodereview.appspot.com/10332220

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@137760 4ff67af0-8c30-449e-8e8b-ad334ec8d88c