BUG=124493
Implementation of a new API, launchAuthFlow, that allows the apps
to perform any auth protocol with any provider.

- Add the API boilerplate (patches from Jon's CL)
- Create a new ExtensionAuthFlow class that encapsulates the
  WebContents / WebView / Widget code to perform the auth flow
- Use ExtensionAuthFlow in the API implementation.

Note that we currently allow two URL formats for redirect back
from the provider:
1. chrome-extension://<extension-id>/
2. https://<extension-id>.ext.chromium.org/

We would ideally want just 1, but providers don't currently allow such a scheme.
So we allow 2 as a workaround and if/when providers add support for 1, we can
remove 2.
Review URL: https://chromiumcodereview.appspot.com/10178020

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@138039 4ff67af0-8c30-449e-8e8b-ad334ec8d88c