[chromedriver] Determine the element location and click synchronously in the renderer.
This is only the browser-side implementation. In this patch we send some scripts from the client which the renderer evaluates to determine where to move or click the mouse. I tried to keep the renderer code as small and as simple as possible. I plan on eventually removing the script-based location finding in favor of some scrollIntoView-like method in WebKit. Currently scrollIntoView does not work. See http://code.google.com/p/chromium/issues/detail?id=73953.

BUG=chromedriver:22,chromedriver:36,chromedriver:44
TEST=none

Review URL: https://chromiumcodereview.appspot.com/10384023

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@138479 4ff67af0-8c30-449e-8e8b-ad334ec8d88c