I've separated out the model from the UI widgets into a separate "data" folder. Some of the classes in the "data" folder are event targets that dispatch events when their state changes. UI widgets listen to these state changes and update their UI accordingly. The two most important "data" classes are PrintTicketStore and DestinationStore. These two manage the state of the print ticket (aka print settings) and the list of all of the print destinations respectively.
The UI widgets are divided into top-level components, settings components, and previewarea components.

Top-level components just include the outer most container widget (called PrintPreview) which handles global events and PrintHeader which renders the print and cancel buttons and some statistics on the document. If you were to start reviewing, PrintPreview is where I would start.

Settings components include all of the UI widgets that can be used to change the state of the PrintTicketStore. They live in the "settings" folder and include "CopiesSettings" and "PageSettings".

Preview area components include all of the UI widgets involved on the main part of the preview UI. PreviewArea and CustomMargins are the two main widgets in the "previewarea" folder.

It's important to note that almost all of the UI widgets have a dependency on PrintTicketStore. Crucially, this allows all UI widgets to be independent from each other and just depend on a common object.

Previously, the state of the print ticket (aka print settings) were distributed in the various UI controls. For example, custom margins were stored both in the preview area and in the margin settings and when it came time to print, these UI components had to be queried for that information, or the question on whether to show the header-footer setting depended on the state of the preview area. Now the state is consolidated into PrintTicketStore. This allows modification of each of the UI widgets without the worry of breaking other widgets.

BUG=114206
TEST=

Review URL: https://chromiumcodereview.appspot.com/10108001

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@138737 4ff67af0-8c30-449e-8e8b-ad334ec8d88c