Add a non-identfying entropy source to MetricsService and use it in FieldTrialList.
We create a new pref, called the low entropy source, which will replace the client ID for use in FieldTrialList for persistence (One Time Randomization). Note that we continue to use the client ID in all other MetricsService activities in the same manner as before.

We also provide a switch that can reset this pref (regenerate the low entropy source value).

Currently we just pass the entropy source directly into the FieldTrialList, which means that all calls to IsOneTimeRandomizationAvailable will always return true. We will have to evaluate what this means for FTL and adjust that in a future CL.

BUG=126546
TEST=Ensure that running Chrome (or Chromium) with UMA disabled and the --reset-low-entropy-source flag causes the user_experience_metrics.low_entropy_source pref to be reset with a new value between 1 and (2^13).


Review URL: https://chromiumcodereview.appspot.com/10335011

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@138817 4ff67af0-8c30-449e-8e8b-ad334ec8d88c