Removed superfluous dcheck.
Removed dcheck from HandleShowSyncSetup() because the promo page likes to call this routine when it's loaded/reloaded. There's no harm in calling this twice since OpenSyncSetup() handles the case when the overlay is already visible.

The other half of this bug is r125797 - if we reload the page, then before the page is reloaded the original SyncSetupHandler should get some kind of a notification so it can unregister itself from LoginUIService.

BUG=129490
TEST=Signin to acct with custom passphrase via promo, when prompted for passphrase hit reload, should not get a crash.

Review URL: https://chromiumcodereview.appspot.com/10453003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@138843 4ff67af0-8c30-449e-8e8b-ad334ec8d88c