gdata: Handle cancellation of the pinning process.
If user cancels download of a pinned-but-not-yet-fetched file, mark
the file as unpinned so that we do not sync the file again upon the
next login.

TODO(hshi): crbug.com/127138 notify when file properties change so
that file manager can clear the "Available offline" checkbox when
the file gets unpinned due to the above reason.

BUG=126742
TEST=lumpy


Review URL: https://chromiumcodereview.appspot.com/10387227

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@139097 4ff67af0-8c30-449e-8e8b-ad334ec8d88c