Persist NaCl's validation cache to disk.
Persisting the cache is straightforward: it is pickled, and the pickle data is sent to the blocking pool to be written to disk.  Reading the cache off disk is a little trickier, because it needs to be done before sel_ldr is started in the NaCl process.  This is implemented as an asynchronous task that blocks starting sel_ldr, similar to opening the IRT.

Functionality to serialize and deserialize the validation cache was also added.  To improve testability, features such as cache enabling / disabling and UMA reporting were shifted from NaClValidationCache onto the NaClBrowser singleton.  This allowed unit tests to be written for NaClValidationCache.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=2515
TEST= Run nexe in browser with NACL_VALIDATION_CACHE=1


Review URL: https://chromiumcodereview.appspot.com/10399053

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@139179 4ff67af0-8c30-449e-8e8b-ad334ec8d88c