Initial landing of experimental wallpaper ui(wallpaper manager).
This experimental wallpaper is disabled by default. To enable it, go to about:flags
and enable "experimental wallpaper UI" or add --experimental-wallpaper-ui to the
command line.

It currently designed as an extension which overrides url chrome://wallpapers. It's
basically the same as file manager and bookmark manager which overrides chrome://files
and chrome://bookmarks.

This is just a place holder for now.


BUG=118684
TEST=


Review URL: https://chromiumcodereview.appspot.com/10391116

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@139265 4ff67af0-8c30-449e-8e8b-ad334ec8d88c