RefCounted types should not have public destructors
Change content::ContentBrowserClient and content::NotificationObserver to
have protected destructors, as their lifetimes are not managed by code
within the content/ code.

Update the few affected callsites to use the most-derived classes, which
have public destructors and for which the callers manage the lifetime of.

BUG=123295
TEST=existing


Review URL: https://chromiumcodereview.appspot.com/10413027

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@139266 4ff67af0-8c30-449e-8e8b-ad334ec8d88c