Add test for DeviceUserWhitelist and DeviceAllowNewUsers policies
This CL adds a PyAuto test for the closely interrelated
DeviceUserWhitelist and DeviceAllowNewUsers policies that control who
can log into a Chrome OS device.
As a prerequisite, the browser automation framework is fixed so that it
returns an error code when login is prevented by the whitelist rather
than timing out.

BUG=129722
TEST=New PyAuto ChromeosDevicePolicy.testUserWhitelistAndAllowNewUsers


Review URL: https://chromiumcodereview.appspot.com/10443028

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@139283 4ff67af0-8c30-449e-8e8b-ad334ec8d88c