Fix for sync case in code forcing extension updates to use https
The previous CL I landed for this bug did not properly handle the case of sync,
where we "update" from a version of 0 of an extension/app to whatever the
current version is.

BUG=129587
TEST=Starting from a clean profile directory, sign in to sync with an account
that has some extensions and/or apps. You should not get any DHECK failure.


Review URL: https://chromiumcodereview.appspot.com/10444063

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@139371 4ff67af0-8c30-449e-8e8b-ad334ec8d88c