Roll libphonenumber r456
This also changes libphonenumber.gyp. A separate library with no metadata needs
to be built. This allows us to build a production library with production
metadata and the unit tests with testing metadata. Before that it was unclear
which metadata was used for testing.

TEST=libphonenumber_unittests


Review URL: https://chromiumcodereview.appspot.com/10412031

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@139410 4ff67af0-8c30-449e-8e8b-ad334ec8d88c