Use HTTPS when responding to user feedback.
This is not a particularly sensitive action, but there are (hopefully
low-sensitivity) cookies in the scope of the origin, and it's a generally
good idea to normalize the use of HTTPS over HTTP (as we have done with e.g.
"report a translation error". This issue was reported to us by Emanuel B
e3amn2l@gmail.com (no bug filed).

TEST=The response page still loads

Review URL: https://chromiumcodereview.appspot.com/10456014

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@139618 4ff67af0-8c30-449e-8e8b-ad334ec8d88c