Fix bug 127498: Panels: The position of notifications aren't adjusted when panels are detached
I fixed a bunch of other issues that could cause balloons not to repositioned in addition to the issue described in the bug. We now observe a new NOTIFICATION_PANEL_STRIP_CHANGED notification that is triggered each time when strip needs to be updated. In addition, we also need to listen to NOTIFICATION_PANEL_CHANGED_EXPANSION_STATE because expansion state change does not update the strip.

More tests are added to cover various kinds of panel operations that could affect positioning of desktop notification balloon. They now live in the new test file.

BUG=127498,118168
TEST=New tests

Review URL: https://chromiumcodereview.appspot.com/10454027

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@139965 4ff67af0-8c30-449e-8e8b-ad334ec8d88c