Pass user gesture indicator for RPH and intents.
We are moving to a policy which will change where we show
RPH and web intents registrations between infobar and page-action
type indicator depending on factors like user gesture, previous
showing of the infobar, and perhaps other considerations. This
change gets that user gesture indicator to the policy logic.

It also incorporates a change to the web intents registration IPC
to pass the service info struct instead of individual members. (Adding
the user gesture went over the param limit, and passing the struct
is a good change regardless.)

R=groby@chromium.org,jam@chromium.org
BUG=119438
TEST=*WebIntent*.*


Review URL: https://chromiumcodereview.appspot.com/10447067

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@139996 4ff67af0-8c30-449e-8e8b-ad334ec8d88c