Remove sync's ModelSafeWorkerRegistrar
The ModelSafeWorkerRegistrar interface was a wrapper around the
SyncBackendRegistrar.  When the sync thread wanted to access the
SyncBackendRegistrar's information regarding currently enabled types or
workers, it would use the ModelSafeWorkerRegistrar interface to do it.

This change removes this implicit inter-thread communication.  Where
necessary, it modifies the SyncBackendHost, SyncManager, SyncScheduler
and the SyncSessionContext to ensure that these classes have access to a
fresh copy of the SyncBackendRegistrar's data whenever it is required.

The most biggest consequence of this patch is that the
SyncSessionContext now maintains a copy of the list of ModelSafeWorkers
and routing info, rather than a pointer to the ModelSafeWorkerRegistrar.
Various functions along the path to CleanupDisabledTypes, Configure and
StartSyncingNormally have been updated to ensure that the latest routing
info is made available to the session context.

Future patches may refactor this code to reduce the amount of duplicated
state.  This work was intentionally left to future patches to reduce the
risk of the current change.  The refactoring should be much easier once
we're convinced that the new, explicit inter-thread communication
mechanisms are no more buggy than the existing code.

BUG=103326
TEST=sync_integration_tests

Review URL: https://chromiumcodereview.appspot.com/10388187

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@140122 4ff67af0-8c30-449e-8e8b-ad334ec8d88c