Guard Against Bad Classifications
Added safety valve so that there will be no attempt to set an attribute on an string beyond its end.
Also, be consistent in using NSInteger by carrying around |nextOffset| as one.
Enhanced the diagnostic output for the AutocompleteMatch::ValidateClassifications function so that the underlying perpetrator of the problem can more easily be identified.

NOTE: This is a partial fix in that the true underlying cause of the bad classification range has not yet been addressed. So this bug will be left open.

BUG=121703
TEST=Visit http://www.hurriyet.com.tr/. Bring up a new tab. Type 's' into the omnibox. No crash means success.
Review URL: https://chromiumcodereview.appspot.com/10537010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@140747 4ff67af0-8c30-449e-8e8b-ad334ec8d88c