Omnibox HistoryQuickProvider: add about:flag to control new scoring
Added an three-value about:flag to control whether the HistoryQuickProvider uses the new scoring function or the old.  The three values are new, old, and automatic.  For now automatic means old but in a later change I'll make automatic assign the status according to a field trial.

BUG=
TEST=by hand


Review URL: https://chromiumcodereview.appspot.com/10532051

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@141059 4ff67af0-8c30-449e-8e8b-ad334ec8d88c