Rewrote the logic for determining which visits should contribute to a suggested page's score.
Since I can't ask georgey what the intent behind the old logic was anymore, I rewrote it so that it makes more sense to me. This should also help speed up the query, as well as remove the chance of an infinite loop in the case of cyclic redirect chains.

BUG=125447
TEST=


Review URL: https://chromiumcodereview.appspot.com/10517012

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@141116 4ff67af0-8c30-449e-8e8b-ad334ec8d88c