Move chromeos::GDataFileSystem::GetCacheDirectoryPath and GetCacheFilePath implementation to GDataCache
This is a part of an effort to move cache related code from GDataFileSystem to GDataCache.
GDataFileSystem::GetCacheDirectoryPath and GetCacheFilePath are still provided to serve a number of existing users, this is a temporary solution.
cache_paths() and kMountedArchiveFileExtension are exposed as a temporary solution.
Tests for these methods are not moved from GDataFileSystemTest in this change, they will be removed and reimplemented once we finish moving other methods and get ready to do more comprehensive test in GDataCacheTest.

BUG=131756
TEST=unit_tests --gtest_filter="GData*"

Review URL: https://chromiumcodereview.appspot.com/10545084

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@141223 4ff67af0-8c30-449e-8e8b-ad334ec8d88c