Fixed a bug in the WaitUntilNavigationCompletes automation command.
Previously the hook would try to use reply_message, despite it being
owned by a scoped_ptr in the NavigationNotificationObserver. This
resulted in a segfault when the page was already loaded, as the
scoped_ptr was deleting the object when the observer was deleted.

BUG=None
TEST=functional/apptest.py


Review URL: https://chromiumcodereview.appspot.com/10535081

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@141278 4ff67af0-8c30-449e-8e8b-ad334ec8d88c