When deciding how to resolve conflicts between sync and local TemplateURLs, if one is safe to autoreplace, remove that one; if both are safe to autoreplace, use the normal method to choose which one "wins", but remove the other rather than uniquifying its keyword.
Also, change the conflict resolution code that conditionally generated SyncChanges "if (!models_associated_)" to do it unconditionally.  The old code used this conditional under the assumption that when being called from ProcessSyncChanges(), these changes would be generated, but this was false.  Hopefully this will help reduce the amount of state inconsistency that may be contributing to bug 130448.

BUG=109295,130448
TEST=none
Review URL: https://chromiumcodereview.appspot.com/10538048

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@141572 4ff67af0-8c30-449e-8e8b-ad334ec8d88c