Refactor the content interface for RequestMediaAccessPermission.
This makes RequestMediaAccessPermission a part of WebContentsDelegate instead
of ContentBrowserClient. The request UI (or lack thereof) needs to be
controlled by the WebContentsDelegate so that platform apps can grant permission
differently than a normal browser. This also seems appropriate because
each request is made on behalf of a particular web contents.

BUG=130216
TEST=navigator.webkitGetUserMedia still works for normal web pages

Review URL: https://chromiumcodereview.appspot.com/10542092

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@141792 4ff67af0-8c30-449e-8e8b-ad334ec8d88c