Do the ReloadLocaleResources() call on the IO thread to avoid synchronization issues.
Followup to https://chromiumcodereview.appspot.com/10543015/
Forced the Reload on the same thread (IO) that was racing with the call.  Since Tony doesn't want us to add synchronization to ResourceBundle from a case that's only exercised in test, we'll go this route to stop the race.

R=tony@chromium.org, estade@chromium.org
BUG=95425
TEST=browser_tests --gtest_filter=WebUIBidiCheckerBrowserTestRTL.*TestSettingsFrameStartup --gtest_also_run_disabled_tests

Review URL: https://chromiumcodereview.appspot.com/10546139

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@141964 4ff67af0-8c30-449e-8e8b-ad334ec8d88c