sync: Refactor per-datatype throttling
This CL pulls the code to track throttled data types out of the sync
session context and into a class meant for that purpose.  This new
class, ThrottledDataTypeTracker, also implements code to notify the
AllStatus object whenever the set of throttled datatypes is changed.

The fact that ThrottledDataTypeTracker, which lives in sync/engine,
references AllStatus caused some problems with DEPS checks.  After a few
iterations during code review, this commit now includes the following
additional changes:
- Move all_status.{cc,h} from sync/internal_api to sync/engine.
- Move the SyncManager::Status inner class out of SyncManager and into
  sync/internal_api/public/engine/sync_status.{cc,h}.  The class has
been renamed to SyncStatus.

This CL does not include code to expose the throttled status on the
chrome://sync page, though it does contain functionality we will use to
implement it in a future commit.

BUG=125065
TEST=

Review URL: https://chromiumcodereview.appspot.com/10454105

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@141992 4ff67af0-8c30-449e-8e8b-ad334ec8d88c