Show throttled data types on about:sync
This change will display a section on the about:sync summary page
listing any currently throttled types.  The section will not be
displayed at all unless at least one data type is throttled.  

The about:sync dump will always contain a subtle indication that data
type throttling reporting is enabled.  We can use this if we're trying
to confirm that a given client has no throttled types, but couldn't be
bothered to look up whether or not it was built before or after this
commit.

BUG=125065
TEST=Very manual:
Temporarily added code to simulate a data type throttling and verified
the type was displayed in about:sync.


Review URL: https://chromiumcodereview.appspot.com/10536158

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@142246 4ff67af0-8c30-449e-8e8b-ad334ec8d88c