Support for initializing filesystem when default download is drive.
When the default download location is a drive folder, we may be required to download a file to drive before the file manager has a chance to authenticate and initialize the fileystem. 
Reset gdata external data in DownloadItem for non-gdata downloads if the user switches to a local path.

BUG=133005,132810
TEST=manual.

Review URL: https://chromiumcodereview.appspot.com/10555013

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@142678 4ff67af0-8c30-449e-8e8b-ad334ec8d88c