Revert 143384, too many uninit reports from Valgrind
-----
If there is a device id use that instead of a new one for re-enrolling.

The original device id is stored in the install attributes and we can
read it from there and reuse it instead of generating a new one.

BUG=chromium-os:26157
TEST=Enroll. Login. Stop UI.Delete /var/lib/whitelist/policy.Start UI. Re-enrollment should still succeed.


Review URL: https://chromiumcodereview.appspot.com/10545083

TBR=pastarmovj@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10638004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@143390 4ff67af0-8c30-449e-8e8b-ad334ec8d88c