Protect access to notification_ui_manager in DesktopNotificationService::Observe
It looks like this has been a potential problem for a while, but was recently exposed because of a change to shutdown ordering. Fortunately the fix is pretty straightforward. I wasn't able to repro the bug, but I confirmed that when ~BrowserPolicyConnector() gets called, g_browser_process->notification_ui_manager() is in fact NULL.

BUG=133833
TEST=Check crash reports, ensure no more crashes in DesktopNotificationService::Observe

Review URL: https://chromiumcodereview.appspot.com/10649002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@143416 4ff67af0-8c30-449e-8e8b-ad334ec8d88c