Fix double deletion of PepperFlashSettingsManager::Core.
The current code has a race condition: When the ref count of Core drops to 0 on the UI thread, a task is posted to the IO thread to delete the object. Before it actually gets executed, however, it is possible that an incoming IPC message causes base::Bind() to be called which increases the ref count. In that case, we will eventually delete the object for more than once.

BUG=132818
TEST=None


Review URL: https://chromiumcodereview.appspot.com/10634021

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@143664 4ff67af0-8c30-449e-8e8b-ad334ec8d88c