chromeos: Fix StartupBrowserCreatorImpl::Launch crash.
The crash seems to be caused by a NULL returned from
browser_to_focus->GetActiveWebContents(). The purpose of that statement is to
focus the app window but browser_to_focus we get from BrowserList::GetLastActive
might not be the one that we are expecting. Getting the app contents explicitly
from OpenApplicationWindow should fix the crash.

BUG=134478
TEST=Verify StartupBrowserCreatorImpl::Launch in issue 134478 no longer happens.


Review URL: https://chromiumcodereview.appspot.com/10662025

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@144228 4ff67af0-8c30-449e-8e8b-ad334ec8d88c