views: Make SimpleMessageBoxViews ref-counted to make sure they get deleted at the right time.
Multiple SimpleMessageBoxViews can show up at the same time. Each of these
starts a nested message-loop. However, these SimpleMessageBoxViews can be
deleted in any order. This creates problems if a box in an inner-loop gets
destroyed before a box in an outer-loop. So to avoid this, ref-counting is
used so that the SimpleMessageBoxViews gets deleted at the right time.

BUG=134471
TEST=manually

Review URL: https://chromiumcodereview.appspot.com/10669032

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@144253 4ff67af0-8c30-449e-8e8b-ad334ec8d88c