Reland: move software_renderling_list.json from deps/ to src/
This is the same CL as crrev.com/144472, except we don't remove the old file in this CL.  We'll do it after a while when every bot picks up the file from new location.  This is to work around some bug that grd change is not picked up immediately.

BUG=134143
TEST=unit_tests, tree
TBR=vangelis
Review URL: https://chromiumcodereview.appspot.com/10712005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@144616 4ff67af0-8c30-449e-8e8b-ad334ec8d88c