Notify CloseFile from Pepper to FileSystem.
Remote filesystems need to know when a plugin finished
updating a local cache file in order to commit the change.
This CL is to wire the notification of file close between
the two components.

Since writable OpenFile is not yet implemented for remote filesystem,
the CL itself should not cause behavior change.

BUG=132236
TEST=browser_tests --gtest_filter='*FileIO*'

Review URL: https://chromiumcodereview.appspot.com/10541113

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@145856 4ff67af0-8c30-449e-8e8b-ad334ec8d88c