Instrument serial API code to make testing easier.
After a good lunchtime discussion, I went off to implement some sort
of overriding functionality in ExtensionFunctionDispatcher, so that I
could tell it to instantiate Function' instead of Function when an
extension/app asked for function. I then discovered
ExtensionFunctionDispatcher::OverrideFunction(), which was pretty
much exactly what I wanted.

Using this facility, added a fake SerialOpenFunction that overrides
a new CreateSerialConnection method to instantiate a
FakeEchoSerialConnection. As its name implies, this class pretends
to echo bytes over a serial port.

The practical outcome of this CL is we test more of our code without
having a physical serial device attached to the machine.

BUG=135656
TEST=this lays the ground work for testing of crrev.com/10759004

Review URL: https://chromiumcodereview.appspot.com/10702138

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@145988 4ff67af0-8c30-449e-8e8b-ad334ec8d88c