chromeos: Fix StartupBrowserCreatorImpl::Launch crash.
Previous fix r144228 does not fix the crash. It seems the crash is more likely
to be caused by user closing the app window in a nested message that runs by
synchronous session restore. Add an WebContentsCloseObserver to monitor the app
contents and skip focusing logic if it is closed.

BUG=134478
TEST=Verify the crash in 134478 no longer happens.


Review URL: https://chromiumcodereview.appspot.com/10700145

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@146027 4ff67af0-8c30-449e-8e8b-ad334ec8d88c