Reland r145993.
Original codereview at http://codereview.chromium.org/10766016/

[Sync] Add location parameter to SyncChange in Sync API.

SyncChanges now require the location they were created as a parameter. This
is then used in the case that the change violates and invariant. We display
the location and type of erroneous change in about:sync, in addition to
disabling the offending datatype.

TBR=stevet@chromium.org, rlarocque@chromium.org
BUG=135108
TEST=unit_tests


Review URL: https://chromiumcodereview.appspot.com/10690128

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@146050 4ff67af0-8c30-449e-8e8b-ad334ec8d88c