Revert 146149 - Don't pass the connector check policy flag along.
If a browser is running, and the service process is running the connector, the
connector will occasionally start up a chromium process to check the connector
enablement policy. The flag to do so would get passed to a running browser, if
present, and there is a path through the policy check code that could cause the
browser to quit.

This patch prevents the policy check flag from being passed on; the started
process will be very short lived and does a silent launch.

BUG=134252
TEST=None


Review URL: https://chromiumcodereview.appspot.com/10666010

TBR=scottbyer@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10703157

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@146248 4ff67af0-8c30-449e-8e8b-ad334ec8d88c