On Windows 8 when a metro app is snapped, all desktop applications receive a resize notification as the work area is reduced. In Chrome the tabstrip paint would fail as the WM_GETTITLEBARINFOEX message fails to return the correct minimize offset.
A fix went in to address this by caching the delta of the minimize button which is then applied to the corresponding
client edge. This only addressed the case when the minimize button offset was returned as 0 from the WM_GETTITLEBARINFOEX 
message. At times this offset is negative. 

Fix is to check whether the minimize button offset is > 0 before applying the cached delta.

BUG=135304
R=sky
Review URL: https://chromiumcodereview.appspot.com/10783027

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@146921 4ff67af0-8c30-449e-8e8b-ad334ec8d88c