gdata: Simplify GDataFileSystem::CopyDocumentToDirectory()
Previously, OnDirectoryChanged() was called when the hosted document
was first copied in the root directory, but this was unnecessary as
the file is soon moved to the destination directory.

This weird behavior was found by adding a new test
TransferFileFromLocalToRemote_HostedDocument in GDataFileSystemTest.
It's good to have a test...

BUG=137526
TEST=out/Release/unit_tests --gtest_filter=GData*

Review URL: https://chromiumcodereview.appspot.com/10778025

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@147020 4ff67af0-8c30-449e-8e8b-ad334ec8d88c