This CL reverts:
140439 - Make OTP sign in on by default.  It can be turned off in about:flags
BUG=81265
TEST=Signing in should ask for OTP instead of ASP.

for the M21 branch.  This is being committed to trunk first so we can test it in canary for a little while before merging to M21.  Once merged, this will be reverted so that OTP sign in becomes the default again in M22.

This CL includes some string changes in the about:flags page which normally would not be allowed for merging into M21, but because this reverts back to string used in M20 this should be fine.

Drover did not work for reverting because there have been other changes since 140439 causing build breaks, so doing the revert manually.

OTP sign is still available in the binary, but needs to be manually enabled in about:flags.


BUG=81265
TEST=Signing in should ask for ASP instead of OTP.


Review URL: https://chromiumcodereview.appspot.com/10800010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@147297 4ff67af0-8c30-449e-8e8b-ad334ec8d88c