Refactored SettingsFrontend and forwarding of calls to the backends.
Added a SettingsNamespaceFrontend interface for namespace frontends, that allows different namespaces to forward StorageCallbacks in their own way.
The LOCAL and SYNC namespaces use an implementation that does the same as before, but the new MANAGED namespace can now be implemented using a ValueStore that loads settings from policies, in the UI thread.

BUG=108992
TEST=Everything works as before, all tests are green


Review URL: https://chromiumcodereview.appspot.com/10784035

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@147464 4ff67af0-8c30-449e-8e8b-ad334ec8d88c