sync: Track validity of about:sync fields
This commit changes the way we initialize the fields which get exported
to about:sync's about tab.  The goal was to pre-declare all the data
which would be exported in a complete snapshot, and to track which of
these fields were actually set.  Unset fields will have a default value
and be specially highlighted on about:sync.  This commit includes the
changes necessary to perform this highlighting.

Also in this commit are a few backend changes that are used to determine
from sync_ui_util which fields are properly initialized and which fields
have been unset by anything other than a default constructor.

The signature for ProfileSyncService::QueryDetailedSyncStatus() has been
modified to return an additional parameter indicating whether or not the
returned status actually came from the backend.

The SyncSesionSnapshot takes a different approach. The object is
immutable and has two constructors. If it is constructed with the
zero-argument constructor it is clearly uninitialized, and will remain
so. This change adds a flag which is set or unset based on which
constructor is used.

Finally, the ConstructAboutInforation function and all its helpers have
been moved out of sync_ui_util.cc and into their own file.

BUG=134927
TEST=


Review URL: https://chromiumcodereview.appspot.com/10696107

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@147531 4ff67af0-8c30-449e-8e8b-ad334ec8d88c