Only slow window animations on shift-click on size control
Our previous approach was to slow layer animations whenever the shift key was held down. This caused problems with shift-clicking on a link to open a page in a new window and also with some shortcuts (like Shift-Esc to open task manager).  I'm reverting to the initial behavior of this feature, which is to only slow down window animations for debugging when you shift-click on a window control.

BUG=133246
TEST=manual, ash_unittests


Review URL: https://chromiumcodereview.appspot.com/10815020

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@147774 4ff67af0-8c30-449e-8e8b-ad334ec8d88c