Use extension IDs instead of Extension instance pointers for app_launcher_handler.cc
Replace the the set of Extension pointers with a set of the extension ids. These ids can be used to retrieve the Extension pointers when needed, ensuring that we get either the current valid pointer, or a NULL pointer if the extension is no longer valid (Instead of using invalid pointers and potentially crashing).


BUG=136094
TEST=
1.click on "Remove from chrome" option by right click on application installed application in this case "Google mail checker".
2.Pop up window will be displayed with "Remove" and "cancel" option.
3.Click on Remove.
4.Immediately after that click on "Chrome webstore icon" in new tab.
5.Chrome opens up the web store (instead of crashing).


Review URL: https://chromiumcodereview.appspot.com/10797010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@148109 4ff67af0-8c30-449e-8e8b-ad334ec8d88c