shell dialogs: Move the linux implementations.
This moves the gtk and kde implementations to a common directory so a search will get all the code when we need to remove the normal gtk port.

This isn't a pure move. Parts of SelectFileDialogImplKDE had to be rewritten to not use BrowserThread.

BUG=134529
TEST=none

Review URL: https://chromiumcodereview.appspot.com/10790083

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@148235 4ff67af0-8c30-449e-8e8b-ad334ec8d88c