Add support for invoking the Windows 8 metro style hung renderer dialog box. This functionality is provided by the newly added class HungRendererDialogMetro.
The actual dialog box is displayed by the metro driver.dll and the functionality to create and display
this is invoked via the exported functions ShowDialogBox and DismissDialogBox. The handlers to be run when
the user clicks on the kill process/wait buttons are passed into the ShowDialogBox function.

Factored out the platform specific portion of the kill process functionality in the static function
HungRendererDialogView::KillRendererProcess which is conditionally compiled for the platform.

BUG=125672
R=sky
Review URL: https://chromiumcodereview.appspot.com/10806079

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@148258 4ff67af0-8c30-449e-8e8b-ad334ec8d88c