Download filename determination refactor (3/3)
Move prompting logic from content/ to chrome/. When determining
the target path for a download, we used to notify content/ of
what the target path was along with a flag indicating whether the
user should be prompted. Then DownloadManager would call into the
DownloadManagerDelegate to request a prompt if necessary. This
patch short circuits the logic so that the
DownloadManagerDelegate would prompt the user as necessary and
only pass the final path to the DownloadItem.

BUG=78085
TEST=unit tests
TBR=sky,mirandac

Review URL: https://chromiumcodereview.appspot.com/10704052

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@148703 4ff67af0-8c30-449e-8e8b-ad334ec8d88c