Implement new slider control and disable the volume control when audio is muted.
Horizontal slider is implemented with image resources, and it changes UI based the on/off state.

Vertical slider is kept in the old way, since I don't have the UI resources for drawing the vertical slider. I wonder if I should remove the vertical slider bar supporting code, or leave it as it which is what I choose to do currently.

BUG=124097
TEST=Audio and brightness slider control should be in new UI, and audio slider bar should dim when muted.
TBR=oshima@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10830027

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@148778 4ff67af0-8c30-449e-8e8b-ad334ec8d88c